requirements.txt
requests
openai
GitPython
python-dotenv
